projects/agentic-ai/beavers-choice-paper/requirements.txt
pandas==2.2.3
typing==3.7.4.3
openai==1.76.0
SQLAlchemy==2.0.40
python-dotenv==1.1.0
smolagents==1.2.0
pydantic==2.5.3
json_repair==0.48.0